requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pynacl>=1.5.0
Pillow>=10.0.0
